Fix/add equipment data verification

+++ frontend/src/shared/tools.tsx
@@ -18,18 +18,20 @@ export const Tools = () => {
   const [toolsItems, setToolsItems] = useState<Equipment[]>([]);
 
   const removeItem = async (id_tool: string) => {
-    const updateItems = toolsItems.filter(tool => tool.id_equipo !== id_tool)
-    try {
-      // Call the backend service to delete the tool
-      const res = await deleteEquipment(id_tool, userToken);
-      const {status, data} = res
-      if (status === 200) {
-        alert(data.message)
-        setToolsItems(updateItems);
-        setToolsData(updateItems);
-      }
-    } catch (error) {
-      console.error("Error deleting equipment:", error);
+    if (confirm('¿Estás seguro que quieres eliminar este equipo?')) {
+        const updateItems = toolsItems.filter(tool => tool.id_equipo !== id_tool)
+        try {
+          // Call the backend service to delete the tool
+          const res = await deleteEquipment(id_tool, userToken);
+          const {status, data} = res
+          if (status === 200) {
+            alert(data.message)
+            setToolsItems(updateItems);
+            setToolsData(updateItems);
+          }
+        } catch (error) {
+          console.error("Error deleting equipment:", error);
+        }
     }
   };
